requirements.txt
face-recognition
streamlit
Pillow
numpy